+++ lib/angen_web/controllers/api/event_json.ex
@@ -6,7 +6,11 @@ defmodule AngenWeb.Api.EventJSON do
   @doc """
   Renders a list of models.
   """
-  def success(_) do
-    %{result: "Event created"}
+  def success(%{count: 0}) do
+    %{result: "No events created"}
+  end
+
+  def success(%{count: count}) do
+    %{result: "Event(s) created", count: count}
   end
 end