+++ lib/angen_web/controllers/api/game_controller.ex
@@ -1,13 +1,11 @@
 defmodule AngenWeb.Api.GameController do
   @moduledoc false
   use AngenWeb, :controller
-  alias AngenWeb.UserAuth
-  alias Angen.TextProtocol.TypeConvertors
 
   action_fallback AngenWeb.FallbackController
 
   @spec create_match(Plug.Conn.t(), map()) :: Plug.Conn.t()
-  def create_match(conn, params) do
+  def create_match(_conn, _params) do
     raise "Not implemented"
   end
 end