+++ test/angen_web/controllers/api/token_controller_test.exs
@@ -34,7 +34,7 @@ defmodule AngenWeb.TokenControllerTest do
       user = user_fixture()
       conn = post(conn, ~p"/api/request_token", %{"id" => user.id, "password" => "password", "user_agent" => "agent"})
       response = json_response(conn, 201)
-      # assert  == %{"result" => "Error", "reason" => "Bad parameters"}
+
       assert response["result"] == "Token"
       assert response["token"]["user_id"] == user.id
       assert Map.has_key?(response["token"], "identifier_code")
@@ -46,16 +46,92 @@ defmodule AngenWeb.TokenControllerTest do
     end
   end
 
+  describe "renewing" do
+    test "bad renewal code", %{conn: conn} do
+      user = user_fixture()
+      token_code = get_api_token_code(user: user)
+      token = Angen.Account.get_user_token_by_identifier(token_code)
+
+      token_list = Angen.Account.list_user_tokens(where: [user_id: user.id])
+      assert token_list == [token]
+
+      # Now attempt renewal
+      conn = put_req_header(conn, "token", token_code)
+      conn = post(conn, ~p"/api/renew_token", %{"renewal" => "123"})
+
+      response = json_response(conn, 500)
+      assert response == %{"result" => "Error", "reason" => "No token"}
+    end
+
+    test "no identifier code", %{conn: conn} do
+      user = user_fixture()
+      token_code = get_api_token_code(user: user)
+      token = Angen.Account.get_user_token_by_identifier(token_code)
+
+      token_list = Angen.Account.list_user_tokens(where: [user_id: user.id])
+      assert token_list == [token]
+
+      # Now attempt renewal
+      conn = post(conn, ~p"/api/renew_token", %{"renewal" => token.renewal_code})
+
+      response = response(conn, 401)
+      assert response == "Unauthorised"
+    end
+
+    test "bad identifier code", %{conn: conn} do
+      user = user_fixture()
+      token_code = get_api_token_code(user: user)
+      token = Angen.Account.get_user_token_by_identifier(token_code)
+
+      token_list = Angen.Account.list_user_tokens(where: [user_id: user.id])
+      assert token_list == [token]
+
+      # Now attempt renewal
+      conn = put_req_header(conn, "token", "123")
+      conn = post(conn, ~p"/api/renew_token", %{"renewal" => token.renewal_code})
+
+      response = response(conn, 401)
+      assert response == "Unauthorised"
+    end
+
+    test "correctly done", %{conn: conn} do
+      user = user_fixture()
+      token_code = get_api_token_code(user: user)
+      token = Angen.Account.get_user_token_by_identifier(token_code)
+
+      token_list = Angen.Account.list_user_tokens(where: [user_id: user.id])
+      assert token_list == [token]
+
+      # Now attempt renewal
+      conn = put_req_header(conn, "token", token_code)
+      conn = post(conn, ~p"/api/renew_token", %{"renewal" => token.renewal_code})
+
+      response = json_response(conn, 201)
+      assert response["result"] == "Token"
+      assert response["token"]["user_id"] == user.id
+      assert Map.has_key?(response["token"], "identifier_code")
+      assert Map.has_key?(response["token"], "renewal_code")
+      assert Map.has_key?(response["token"], "expires_at")
+
+      token = Angen.Account.get_user_token_by_identifier(response["token"]["identifier_code"])
+      assert token.user_id == user.id
+
+      # Ensure old one is deleted
+      old_token = Angen.Account.get_user_token_by_identifier(token_code)
+      assert old_token == nil
+    end
+  end
+
   describe "using" do
     test "no token", %{conn: conn} do
-      conn = post(conn, ~p"/api/events/simple_clientapp", %{"name" => "no-token"})
+      conn = post(conn, ~p"/api/events/simple_clientapp", %{"events" => [%{"name" => "no-token"}]})
       response = response(conn, 401)
       assert response == "Unauthorised"
     end
 
     test "bad token", %{conn: conn} do
       conn = put_req_header(conn, "token", "123")
-      conn = post(conn, ~p"/api/events/simple_clientapp", %{"name" => "bad-token"})
+      conn = post(conn, ~p"/api/events/simple_clientapp", %{"events" => [%{"name" => "bad-token"}]})
       response = response(conn, 401)
       assert response == "Unauthorised"
     end
@@ -64,9 +140,9 @@ defmodule AngenWeb.TokenControllerTest do
       token = get_api_token_code()
       conn = put_req_header(conn, "token", token)
 
-      conn = post(conn, ~p"/api/events/simple_clientapp", %{"name" => "good-token"})
+      conn = post(conn, ~p"/api/events/simple_clientapp", %{"events" => [%{"name" => "good-token"}]})
       response = json_response(conn, 201)
-      assert response == %{"result" => "Event created"}
+      assert response == %{"result" => "Event(s) created", "count" => 1}
     end
   end
 end